move name.py from ermpath/ back to url/ast
These name-resolution rules are part of the REST interface and really
didn't belong in ermpath, where they'd been relocated as part of a previous
hack...
